+++ R/recode.R
@@ -35,6 +35,10 @@ fct_recode <- function(.f, ...) {
     new_levels <- new_levels[!nulls]
   }
 
+  lvls_revalue(f, lvls_rename(f, new_levels))
+}
+
+lvls_rename <- function(f, new_levels) {
   # Match old levels with new levels
   old_levels <- levels(f)
   idx <- match(new_levels, old_levels)
@@ -49,8 +53,7 @@ fct_recode <- function(.f, ...) {
   }
 
   old_levels[idx] <- names(new_levels)
-
-  lvls_revalue(f, old_levels)
+  old_levels
 }
 
 check_recode_levels <- function(..., call = caller_env()) {