+++ tests/testthat/test-count.R
@@ -6,13 +6,24 @@ test_that("0 count for empty levels", {
   expect_equal(fct_count(f)$n, c(1, 0, 0))
 })
 
-test_that("counts NA even when not in levels", {
-  f <- factor(c("a", "a", NA))
-  out <- fct_count(f)
+test_that("counts NA values and levels", {
+  f1 <- factor(c("a", "a", NA))
+  expect_equal(
+    fct_count(f1),
+    tibble::tibble(
+      f = fct(c("a", NA)),
+      n = c(2, 1)
+    )
+  )
 
-  expect_equal(out$n, c(2, 1))
-  # and doesn't change levels
-  expect_equal(levels(out$f), levels(f))
+  f2 <- factor(c("a", "a", NA), exclude = NULL)
+  expect_equal(
+    fct_count(f2),
+    tibble::tibble(
+      f = fct(c("a", NA), c("a", NA)),
+      n = c(2, 1)
+    )
+  )
 })
 
 test_that("returns marginal table", {